Glider Audio + Audiomanager upgrade and gameobject finder functionality
Glider sounds added for opening/closing/crashing (also a variable woosh sound that changes pitch+volume in response to velocity). Upgraded the audiomanager to properly work with initially disabled gameobjects + objects created after the audiomanager. This required a class that extends the functionality of the regular game object find functions